+++ packages/bluesky_text/lib/bluesky_text.dart
@@ -7,5 +7,7 @@ export 'package:bluesky_text/src/config/link_config.dart';
 export 'package:bluesky_text/src/entities/byte_indices.dart';
 export 'package:bluesky_text/src/entities/entities.dart';
 export 'package:bluesky_text/src/entities/entity.dart';
+export 'package:bluesky_text/src/text_length_overflow.dart';
+export 'package:bluesky_text/src/text_segment.dart';
 export 'package:bluesky_text/src/utils.dart'
     show isEmojiOnly, getGraphemeLength;